refactor: 비즈니스 로직을 다루는 클래스 domain 패키지로 이동
