+++ src/main/java/racingcar/domain/MovingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface MovingStrategy {
     boolean isMove();
