+++ src/main/java/racingcar/domain/RandomNumberMovingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import racingcar.util.RandomGenerator;
